// app/types/pocketbase-types.ts
export enum Collections {
	Authorigins = "_authOrigins",
	Externalauths = "_externalAuths",
	Mfas = "_mfas",
	Otps = "_otps",
	Superusers = "_superusers",
	Episodes = "episodes",
	Speakers = "speakers",
	Users = "users",
}

// Alias types for improved usability
export type IsoDateString = string
export type RecordIdString = string
export type HTMLString = string

// System fields
export type BaseSystemFields<T = never> = {
	id: RecordIdString
	collectionId: string
	collectionName: Collections
	expand?: T
}
// ...
export type EpisodesRecord = {
	content: HTMLString
	created?: IsoDateString
	date: IsoDateString
	description?: string
	id: string
	image: string
	speakers?: RecordIdString[]
	spotify: string
	title: string
	updated?: IsoDateString
	youtube: string
}

export type SpeakersRecord = {
	bio?: HTMLString
	created?: IsoDateString
	episodes?: RecordIdString[]
	facebook?: string
	id: string
	image?: string
	instagram?: string
	linkedin?: string
	name: string
	nickname?: string
	updated?: IsoDateString
	website?: string
	x?: string
	youtube?: string
}
// ...
export type EpisodesResponse<Texpand = unknown> = Required<EpisodesRecord> & BaseSystemFields<Texpand>
export type SpeakersResponse<Texpand = unknown> = Required<SpeakersRecord> & BaseSystemFields<Texpand>